+++ api/src/Fixture/Category.php
@@ -24,16 +24,6 @@ trait Category
      */
     public function load(ObjectManager $manager)
     {
-        $connection = $manager->getConnection();
-        $platform = $connection->getDatabasePlatform()->getName();
-
-        switch ($platform) {
-            case 'postgresql':
-                $connection->exec('ALTER SEQUENCE app_category_id_seq RESTART WITH 1');
-                $connection->exec('ALTER SEQUENCE app_category_trans_id_seq RESTART WITH 1');
-                break;
-        }
-
         $objects = $this->parse($this->path);
 
         foreach ($objects as $object) {